+++ apt/tools/gensrclist.cc
@@ -31,7 +31,7 @@
 using namespace std;
 
 int tags[] =  {
-   RPMTAG_NAME,
+       RPMTAG_NAME,
        RPMTAG_EPOCH,
        RPMTAG_VERSION,
        RPMTAG_RELEASE,
@@ -44,6 +44,7 @@ int tags[] =  {
        
        RPMTAG_DESCRIPTION, 
        RPMTAG_SUMMARY, 
+       /*RPMTAG_HEADERI18NTABLE*/ HEADER_I18NTABLE,
        
        RPMTAG_REQUIREFLAGS, 
        RPMTAG_REQUIRENAME,
@@ -124,6 +125,14 @@ void usage()
    cerr << "                are in the same directory level"<<endl;
 }
 
+#ifdef HAVE_RPM4
+extern "C" {
+// No prototype from rpm after 4.0.
+int headerGetRawEntry(Header h, int_32 tag, int_32 * type,
+		      void *p, int_32 *c);
+}
+#endif
+
 int main(int argc, char ** argv) 
 {
    char buf[300];
@@ -141,8 +150,10 @@ int main(int argc, char ** argv)
    bool progressBar = false;
    bool flatStructure = false;
    char *arg_dir, *arg_suffix, *arg_srpmindex;
+   const char *srcListSuffix = NULL;
+   bool srcListAppend = false;
 
-   putenv("LC_ALL=");
+   putenv("LC_ALL="); // Is this necessary yet (after i18n was supported)?
    for (i = 1; i < argc; i++) {
       if (strcmp(argv[i], "--mapi") == 0) {
 	 mapi = true;
@@ -150,6 +161,16 @@ int main(int argc, char ** argv)
 	 flatStructure = true;
       } else if (strcmp(argv[i], "--progress") == 0) {
 	 progressBar = true;
+      } else if (strcmp(argv[i], "--append") == 0) {
+	 srcListAppend = true;
+      } else if (strcmp(argv[i], "--meta") == 0) {
+	 i++;
+	 if (i < argc) {
+	    srcListSuffix = argv[i];
+	 } else {
+	    cout << "gensrclist: argument missing for option --meta"<<endl;
+	    exit(1);
+	 }
       } else {
 	 break;
       }
@@ -220,11 +241,17 @@ int main(int argc, char ** argv)
 
    chdir(buf);
    
-   sprintf(buf, "%s/srclist.%s", cwd, arg_suffix);
+   if (srcListSuffix != NULL)
+      sprintf(buf, "%s/srclist.%s", cwd, srcListSuffix);
+   else
+      sprintf(buf, "%s/srclist.%s", cwd, arg_suffix);
    
-   unlink(buf);
-   
-   outfd = fdOpen(buf, O_WRONLY | O_TRUNC | O_CREAT, 0644);
+   if (srcListAppend == true && FileExists(buf)) {
+      outfd = fdOpen(buf, O_WRONLY|O_APPEND, 0644);
+   } else {
+      unlink(buf);
+      outfd = fdOpen(buf, O_WRONLY|O_TRUNC|O_CREAT, 0644);
+   }
    if (!outfd) {
       cerr << "gensrclist: error creating file" << buf << ":"
 	  << strerror(errno);
@@ -283,14 +310,11 @@ int main(int argc, char ** argv)
 	       void *data;
 	       int res;
 	       
-	       res = headerGetEntry(h, tags[i], &type, &data, &count);
-	       if (res != 1) {
-		  /*
-		   printf("warning: tag %i not found on header for %s\n",
-		   tags[i], dirEntries[entry_cur]->d_name);
-		   */
+	       // Copy raw entry, so that internationalized strings
+	       // will get copied correctly.
+	       res = headerGetRawEntry(h, tags[i], &type, &data, &count);
+	       if (res != 1)
 		  continue;
-	       }
 	       headerAddEntry(newHeader, tags[i], type, data, count);
 	    }
